angle2pixels {gazetools}	R Documentation

Angle to Pixels

Description

Returns a distance (in pixels) given a visual angle and distance.

Usage

  angle2pixels(alpha, rx, ry, sw, sh, ez, ex = 0, ey = 0,
    x = rx/2, y = ry/2)

Arguments

alpha
the target visual angle

rx
the x resolution of the monitor (pixels)

ry
the y resolution of the monitor (pixels)

sw
the physical screen width (mm)

sh
the physical screen height (mm)

ez
the perpendicular distance from the viewer to the screen (mm)

ex
the horizontal offset of the viewer from screen center (mm)

ey
the vertical offset of the viewer from screen center (mm)

x
the x coordinate of a point on a screen (pixels)

y
the y coordinate of a point on a screen (pixels)

[Package gazetools version 3.0 Index]