chull_area {gazetools}	R Documentation

Convex Hull Area

Description

Computes the area of minimal convex polygon that spans all fixations.

Usage

  chull_area(x)

Arguments

x
coordinate vectors of points. This can be specified as a 2-column matrix x, a list x with two components

[Package gazetools version 3.0 Index]