classify.VI {gazetools}	R Documentation

Gaze Data Classification (velocity-iterative)

Description

Classifies gaze data into saccades and fixations using an iteratively determined velocity threshold.

Usage

  classify.VI(v, vt = 100, sigma = 6)

Arguments

v
a vector of instantaneous velocity values

vt
a velocity threshold

sigma
the number of standard deviations between fixation and saccade velocities

[Package gazetools version 3.0 Index]