distance2point {gazetools}	R Documentation

Distance to Point

Description

Takes an x and y screen coordinate and returns the physical distance from the observer to that point on the screen.

Usage

  distance2point(x, y, rx, ry, sw, sh, ez, ex = 0, ey = 0)

Arguments

x
the x coordinate of a point on a screen (pixels)

y
the y coordinate of a point on a screen (pixels)

rx
the x resolution of the monitor (pixels)

ry
the y resolution of the monitor (pixels)

sw
the physical screen width (mm)

sh
the physical screen height (mm)

ez
the perpendicular distance from the viewer to the screen (mm)

ex
the horizontal offset of the viewer from screen center (mm)

ey
the vertical offset of the viewer from screen center (mm)

[Package gazetools version 3.0 Index]