pva {gazetools}	R Documentation

Position, Velocity, Acceleration

Description

Takes standard eye tracker gaze information and computes smoothed position, velocity and acceleration profiles using a Savitzky-Golay filter.

Usage

  pva(x, y, time, samplerate, rx, ry, sw, sh, ez, ex = 0,
    ey = 0, order = 2, window = 11)

Arguments

x
the x coordinate of a point on a screen (pixels)

y
the y coordinate of a point on a screen (pixels)

time
the time corresponding to raw gaze samples

samplerate
the samplerate of the eyetracker

rx
the x resolution of the monitor (pixels)

ry
the y resolution of the monitor (pixels)

sw
the physical screen width (mm)

sh
the physical screen height (mm)

ez
the perpendicular distance from the viewer to the screen (mm)

ex
the horizontal offset of the viewer from screen center (mm)

ey
the vertical offset of the viewer from screen center (mm)

order
the Savitzky-Golay filter order

window
the Savitzky-Golay filter window length (must be odd)

[Package gazetools version 3.0 Index]