pva-class {gazetools}	R Documentation

Class "pva"

Description

A class to hold position, velocity and acceleration of raw gaze data

Slots

time:

vector of class "numeric", containing the time corresponding to raw gaze samples

ez:

vector of class "numeric", containing the perpendicular distance from the viewer to the screen (mm)

ex:

vector of class "numeric", containing the horizontal offset of the viewer from screen center (mm)

ey:

vector of class "numeric", containing the vertical offset of the viewer from screen center (mm)

x:

vector of class "numeric", containing the x coordinate of a point on a screen (pixels)

y:

vector of class "numeric", containing the y coordinate of a point on a screen (pixels)

sx:

vector of class "numeric", containing the smoothed x coordinate of a point on a screen (pixels)

sy:

vector of class "numeric", containing the smoothed y coordinate of a point on a screen (pixels)

xa:

vector of class "numeric", containing the x coordinate of a point in visual angle relative to the center of screen (degrees)

ya:

vector of class "numeric", containing the y coordinate of a point in visual angle relative to the center of screen (degrees)

v:

vector of class "numeric", containing the instantaneous velocity (degrees/s)

a:

vector of class "numeric", containing the instantaneous acceleration (degrees/s^2)

rx:

the x resolution of the monitor (pixels)

ry:

the y resolution of the monitor (pixels)

samplerate:

the samplerate of the eyetracker

sgolayfilt:

parameters for the avitzky-Golay smoothing filter

[Package gazetools version 3.0 Index]