mouldThreshold {gazetools}	R Documentation

Mould Threshold

Description

Uses the Mould algorithm to determine the optimal velocity or acceleration threshold for classifying raw gaze data.

Arguments

dpva
an object of class "pva"

type
'velocity' or 'acceleration'

[Package gazetools version 3.0 Index]