subtendedAngle {gazetools}	R Documentation

Subtended Angle

Description

Takes two screen coordinates and returns the angle (in degrees) subtended by those two points.

Usage

  subtendedAngle(x1, y1, x2, y2, rx, ry, sw, sh, ez,
    ex = 0, ey = 0)

Arguments

x1
the x coordinate of the first point on a screen (pixels)

y1
the y coordinate of the first point on a screen (pixels)

x2
the x coordinate of the second point on a screen (pixels)

y2
the y coordinate of the second point on a screen (pixels)

rx
the x resolution of the monitor (pixels)

ry
the y resolution of the monitor (pixels)

sw
the physical screen width (mm)

sh
the physical screen height (mm)

ez
the perpendicular distance from the viewer to the screen (mm)

ex
the horizontal offset of the viewer from screen center (mm)

ey
the vertical offset of the viewer from screen center (mm)

[Package gazetools version 3.0 Index]