voronoi_polygons {gazetools}	R Documentation

Voronoi Polygons

Description

Computes Voronoi polygons based on a set of x,y coordinates

Usage

  voronoi_polygons(x, rw)

Arguments

x
a matrix of x,y cooridinates

rw
the coordinates of the corners of the rectangular window enclosing the voronoi cells, in the order (xmin, xmax, ymin, ymax).

[Package gazetools version 3.0 Index]