voronoi_skewness {gazetools}	R Documentation

Voronoi Skewness

Description

Given a matrix of x,y coordinates computes the skewness of the area of the voronoi cells

Usage

  voronoi_skewness(d, rw)

Arguments

d
a matrix of x,y coordinates

rw
the coordinates of the corners of the rectangular window enclosing the voronoi cells, in the order (xmin, xmax, ymin, ymax).

[Package gazetools version 3.0 Index]